# Conflicts:
#	dev-tools/scripts/smokeTestRelease.py

+++ solr/solrj/src/java/org/apache/solr/common/util/FastJavaBinDecoder.java
@@ -285,7 +285,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
     void reset() {
       this.doubleVal = 0.0d;
-      this.numericVal = 0l;
+      this.numericVal = 0L;
       this.objVal = null;
       this.ctx = null;
       this.entryListener = null;
@@ -338,7 +338,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Byte.valueOf((byte) entry.numericVal);
+        return (byte) entry.numericVal;
       }
     },
     _SHORT(SHORT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -350,7 +350,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Short.valueOf((short) entry.numericVal);
+        return (short) entry.numericVal;
       }
     },
     _DOUBLE(DOUBLE, LOWER_5_BITS, DataEntry.Type.DOUBLE) {
@@ -362,7 +362,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Double.valueOf(entry.doubleVal);
+        return entry.doubleVal;
       }
     },
     _INT(INT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -373,7 +373,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Integer.valueOf((int) entry.numericVal);
+        return (int) entry.numericVal;
       }
     }, // signed integer
     _LONG(LONG, LOWER_5_BITS, DataEntry.Type.LONG) {
@@ -384,7 +384,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Long.valueOf(entry.numericVal);
+        return entry.numericVal;
       }
     },
     _FLOAT(FLOAT, LOWER_5_BITS, DataEntry.Type.FLOAT) {
@@ -395,7 +395,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return Float.valueOf((float) entry.doubleVal);
+        return (float) entry.doubleVal;
       }
     },
     _DATE(DATE, LOWER_5_BITS, DataEntry.Type.DATE) {
